+++ options/cte.go
@@ -38,7 +38,7 @@ type CTEDecoderOptions struct {
 	BufferSize int
 
 	// Concise encoding spec version to adhere to.
-	ConciseEncodingVersion uint
+	ConciseEncodingVersion uint64
 
 	// The implied structure that this decoder will assume.
 	// Any implied structure will be automatically reported without being
@@ -74,7 +74,7 @@ type CTEEncoderOptions struct {
 	BinaryFloatEncoding BinaryFloatEncodeAs
 
 	// Concise encoding spec version to adhere to.
-	ConciseEncodingVersion uint
+	ConciseEncodingVersion uint64
 
 	// The implied structure that this encoder will assume.
 	// Any implied structure will not actually be written to the document.
